requirements.txt
pandas
yfinance
sqlalchemy
psycopg2
requests
pandas_market_calendars
apache-airflow